Refactor User instantiation using reflection
Updated the code to create an instance of the `User` class dynamically using reflection. Replaced all references to the old `user` variable with the new `userInstance`. Adjusted method invocations and field accesses accordingly. Changed the age parameter in `ChangeAge` from `100` to `20` for specific logic adjustment. These changes improve the flexibility of the code.